+++ appdaemon/adapi.py
@@ -2463,7 +2463,8 @@ class ADAPI:
                 It must conform to the standard Scheduler Callback format documented
                 `here <APPGUIDE.html#about-schedule-callbacks>`__.
             start: A Python ``datetime`` object that specifies when the initial callback
-                will occur.
+                will occur, or can take the `now` string alongside an added offset. If given
+                in the past, it will be executed in the next interval time.
             interval: Frequency (expressed in seconds) in which the callback should be executed.
             **kwargs: Arbitrary keyword parameters to be provided to the callback
                 function when it is invoked.
@@ -2488,13 +2489,32 @@ class ADAPI:
             Run every 17 minutes starting in 2 hours time.
 
             >>> self.run_every(self.run_every_c, time, 17 * 60)
+            
+            Run every 10 minutes starting now.
+
+            >>> self.run_every(self.run_every_c, "now", 10 * 60)
+            
+            Run every 5 minutes starting now plus 5 seconds.
+
+            >>> self.run_every(self.run_every_c, "now+5", 5 * 60)
 
         """
         name = self.name
         now = await self.get_now()
-        aware_start = self.AD.sched.convert_naive(start)
+        
+        if isinstance(start, str) and "now" in start: # meaning immediate time required
+            now_offset = 0
+            if "+" in start: # meaning time to be added
+                now_offset = int(re.findall(r"\d+", start)[0])
+            
+            aware_start = await self.get_now()
+            aware_start = aware_start + datetime.timedelta(seconds=now_offset)
+            
+        else:
+            aware_start = self.AD.sched.convert_naive(start)
+
         if aware_start < now:
-            raise ValueError("start cannot be in the past")
+            aware_start = now + datetime.timedelta(seconds=interval)
 
         self.logger.debug(
             "Registering run_every starting %s in %ss intervals for %s", aware_start, interval, name,